--- a/testdata/bench_after/Connected
+++ b/testdata/bench_after/Connected
@@ -1,0 +1,7 @@
+goos: linux
+goarch: amd64
+pkg: github.com/rschio/graph
+cpu: Intel(R) Core(TM) i5-8350U CPU @ 1.70GHz
+BenchmarkConnected-8   	   11476	    103912 ns/op	    8267 B/op	       5 allocs/op
+PASS
+ok  	github.com/rschio/graph	1.986s