--- a/testdata/bench_after/EulerDirected
+++ b/testdata/bench_after/EulerDirected
@@ -1,0 +1,7 @@
+goos: linux
+goarch: amd64
+pkg: github.com/rschio/graph
+cpu: Intel(R) Core(TM) i5-8350U CPU @ 1.70GHz
+BenchmarkEulerDirected-8   	  589666	      2005 ns/op	     960 B/op	       4 allocs/op
+PASS
+ok  	github.com/rschio/graph	2.041s