--- a/testdata/bench_after/EulerUndirected
+++ b/testdata/bench_after/EulerUndirected
@@ -1,0 +1,7 @@
+goos: linux
+goarch: amd64
+pkg: github.com/rschio/graph
+cpu: Intel(R) Core(TM) i5-8350U CPU @ 1.70GHz
+BenchmarkEulerUndirected-8   	  508767	      2387 ns/op	     968 B/op	       5 allocs/op
+PASS
+ok  	github.com/rschio/graph	2.188s