--- a/testdata/bench_after/MST
+++ b/testdata/bench_after/MST
@@ -1,0 +1,7 @@
+goos: linux
+goarch: amd64
+pkg: github.com/rschio/graph
+cpu: Intel(R) Core(TM) i5-8350U CPU @ 1.70GHz
+BenchmarkMST-8   	    5379	    189717 ns/op	   32951 B/op	       8 allocs/op
+PASS
+ok  	github.com/rschio/graph	1.048s