--- a/testdata/bench_after/MaxFlow
+++ b/testdata/bench_after/MaxFlow
@@ -2,6 +2,6 @@ goos: linux
 goarch: amd64
 pkg: github.com/rschio/graph
 cpu: Intel(R) Core(TM) i5-8350U CPU @ 1.70GHz
-BenchmarkMaxFlow-8   	       1	2062536970 ns/op	15380024 B/op	   21500 allocs/op
+BenchmarkMaxFlow-8   	       1	2086161963 ns/op	15496176 B/op	   21334 allocs/op
 PASS
-ok  	github.com/rschio/graph	2.087s
+ok  	github.com/rschio/graph	2.108s