--- a/testdata/bench_before/EulerDirected
+++ b/testdata/bench_before/EulerDirected
@@ -1,0 +1,7 @@
+goos: linux
+goarch: amd64
+pkg: github.com/rschio/graph
+cpu: Intel(R) Core(TM) i5-8350U CPU @ 1.70GHz
+BenchmarkEulerDirected-8   	  147134	      7276 ns/op	    5712 B/op	     103 allocs/op
+PASS
+ok  	github.com/rschio/graph	1.159s