--- a/testdata/bench_before/EulerUndirected
+++ b/testdata/bench_before/EulerUndirected
@@ -1,0 +1,7 @@
+goos: linux
+goarch: amd64
+pkg: github.com/rschio/graph
+cpu: Intel(R) Core(TM) i5-8350U CPU @ 1.70GHz
+BenchmarkEulerUndirected-8   	  152704	      7065 ns/op	    5720 B/op	     104 allocs/op
+PASS
+ok  	github.com/rschio/graph	1.164s