--- a/testdata/bench_before/Sort
+++ b/testdata/bench_before/Sort
@@ -1,0 +1,7 @@
+goos: linux
+goarch: amd64
+pkg: github.com/rschio/graph
+cpu: Intel(R) Core(TM) i5-8350U CPU @ 1.70GHz
+BenchmarkSort-8   	    2175	    608159 ns/op	  249174 B/op	    6030 allocs/op
+PASS
+ok  	github.com/rschio/graph	1.383s