--- a/testdata/bench_before/TopSort
+++ b/testdata/bench_before/TopSort
@@ -1,0 +1,7 @@
+goos: linux
+goarch: amd64
+pkg: github.com/rschio/graph
+cpu: Intel(R) Core(TM) i5-8350U CPU @ 1.70GHz
+BenchmarkTopSort-8   	    9980	    134713 ns/op	   56076 B/op	    1180 allocs/op
+PASS
+ok  	github.com/rschio/graph	2.332s