--- a/testdata/bench_before/Transpose
+++ b/testdata/bench_before/Transpose
@@ -1,0 +1,7 @@
+goos: linux
+goarch: amd64
+pkg: github.com/rschio/graph
+cpu: Intel(R) Core(TM) i5-8350U CPU @ 1.70GHz
+BenchmarkTranspose-8   	    2698	    404269 ns/op	  217577 B/op	    5001 allocs/op
+PASS
+ok  	github.com/rschio/graph	1.140s