--- a/testdata/bench_results/Acyclic
+++ b/testdata/bench_results/Acyclic
@@ -1,0 +1,8 @@
+name       old time/op    new time/op    delta
+Acyclic-8     128µs ± 0%      66µs ± 0%  -47.95%
+
+name       old alloc/op   new alloc/op   delta
+Acyclic-8    52.0kB ± 0%    12.4kB ± 0%  -76.18%
+
+name       old allocs/op  new allocs/op  delta
+Acyclic-8     1.17k ± 0%     0.01k ± 0%  -98.89%