… path_test.go
… top_test.go
… weak_test.go